Remove more public headers
These aren't needed anymore.
