--- sky/engine/platform/exported/linux/WebFontRenderStyle.cpp
@@ -1,61 +1,0 @@
-/*
- * Copyright (C) 2010 Google Inc. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are
- * met:
- *
- *     * Redistributions of source code must retain the above copyright
- * notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above
- * copyright notice, this list of conditions and the following disclaimer
- * in the documentation and/or other materials provided with the
- * distribution.
- *     * Neither the name of Google Inc. nor the names of its
- * contributors may be used to endorse or promote products derived from
- * this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#include "flutter/sky/engine/public/platform/linux/WebFontRenderStyle.h"
-
-#include "flutter/sky/engine/platform/fonts/FontRenderStyle.h"
-
-using blink::FontRenderStyle;
-
-namespace blink {
-
-void WebFontRenderStyle::toFontRenderStyle(FontRenderStyle* out)
-{
-    out->useBitmaps = useBitmaps;
-    out->useAutoHint = useAutoHint;
-    out->useHinting = useHinting;
-    out->hintStyle = hintStyle;
-    out->useAntiAlias = useAntiAlias;
-    out->useSubpixelRendering = useSubpixelRendering;
-    out->useSubpixelPositioning = useSubpixelPositioning;
-}
-
-void WebFontRenderStyle::setDefaults()
-{
-    useBitmaps = 2;
-    useAutoHint = 2;
-    useHinting = 2;
-    hintStyle = 0;
-    useAntiAlias = 2;
-    useSubpixelRendering = 2;
-    useSubpixelPositioning = 2;
-}
-
-} // namespace blink
